refactor(ogr): bring OGR Metrics outliers under intentional limits
Replace OGR::Geometry.factory's large WKB-type case with a frozen
type-to-class-name map resolved lazily via const_get (the geometry subclasses
load after this file), preserving the wkbLineString-to-LinearRing/LineString
special case and the passthrough fallback. Move the OGR::Geometry::ClassMethods
module into its own file for clarity.

Collapse the duplicated DateTime construction in OGR::Feature#field_as_date_time
into a single call, and extract the extent-ring construction in
OGR::Layer#geometry_from_extent into a private helper.

Document with scoped disables the cohesive collections that would only be
fragmented by splitting: OGR::Geometry's flat set of thin OGR_G_* wrappers,
OGR::Feature's field accessors, the OSR coordinate-system getters/setters, and
OGR::Layer#point_values (kept inline for gridding performance, per its note).
